Add Rakefile to build guarded files
Create a basic Rakefile that just requests that guard build the files
that it knows about.

This isn't an optimal solution because failure of guard to build any
files will not prevent rake from proceeding with any tasks that depend
on those files, but the other option would be to duplicate build rules
between guard and rake which is a bit uglier.