Don't flag whitespace at EOL as errors in markdown
Add a .gitattributes file specifying git shouldn't highlight whitespace
at end of line as an error.  I have vim configured to include the `w`
flag in the `formatoptions` setting which uses trailing whitespace on a
line to indicate that the paragraph continues onto the next line.  This
allows it to reflow paragraphs automatically with less likelihood of
merging lines that where the break is important.